+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v 9.3.1
 
+- new: post on website in sharing menu
 - fix: Half screen after orientation change #919
 - fix: not all menu items accessible in landscape orientation
 - fix: Download dialog does not decode multibyte file name #917